Split handlers.go
